refactor(api-gateway): Phase 10 reshape ApiGateway.Tests into Features/ + Infrastructure/ + Composition/
- Move per-endpoint operator tests under ApiGateway.Tests/Features/Operator/{GetFailureDetail,DiscardFailure,BatchReplayFailures}/; cross-slice routing + DI boot tests under Composition/; OutboxFailurePollerTests under Infrastructure/Polling/; proxy + Swagger-transformer tests under Infrastructure/Proxy/[/SwaggerAggregation/]. Namespaces updated; behavior unchanged.
- Delete now-empty Operator/, Gateway/, Gateway/SwaggerAggregation/ test folders.
- ListFailures/ + ReplayFailure/ slice test folders not populated: no per-endpoint test class exists today; the issue text permits this (those routes remain covered by cross-slice OperatorEndpointsRoutingTests in Composition/).
